clientv3: disable mirror auth test with proxy
